The functions are generated for the three possible combinations of inNotSyncZone and inParallelZone
It looks like we have a stable version for not_sync and parallel_begin_end
and their integration. The new .tpl file shows that. You can use it
with the tplToExe.sh script.

[The order of the function generation its solved for the main() case
but still the order of the functions in the cpp file depends on the ordering
of the names in the function HashMap data structure.]